requirements.txt
bitarray ~= 2.6
humanfriendly ~= 10.0
pyfarmhash ~= 0.3.2
xxhash ~= 3.0.0
